remove all mentions to color-scheme old name and replacing them with cupertino
