+++ colors/cupertino.lua
@@ -1,0 +1,1 @@
+require("cupertino").colorscheme()
