+++ lua/cupertino/config.lua
